pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.22.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-docx>=1.2.0",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
]
